requirements.txt
anyascii==0.3.1
appdirs==1.4.4
APScheduler==3.9.1
apturl==0.5.2
audioread==2.1.9
autopep8==1.6.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.0.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cmudict==1.0.2
colorama==0.4.4
command-not-found==0.3
contractions==0.1.72
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
Flask==2.1.2
Flask-Cors==3.0.10
future==0.18.2
gevent==21.12.0
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-datastore==2.5.1
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.46.0
grpcio-status==1.46.0
gunicorn==20.1.0
httplib2==0.20.2
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.6.4
inflect==5.6.0
inotify==0.2.10
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.1.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
librosa==0.9.1
llvmlite==0.38.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
nltk==3.7
nose==1.3.7
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.4.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pooch==1.6.0
praw==7.6.0
prawcore==2.3.0
pronouncing==0.2.0
proto-plus==1.20.3
protobuf==3.20.1
ptyprocess==0.7.0
pyahocorasick==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytrends==4.8.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
rake-nltk==1.0.6
regex==2022.4.24
reportlab==3.6.8
requests==2.25.1
resampy==0.2.2
rsa==4.8
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
SoundFile==0.10.3.post1
systemd-python==234
textsearch==0.0.21
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
transformers==4.18.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
update-checker==0.18.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
websocket-client==1.3.2
Werkzeug==2.1.2
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
